+++ includes/class-testimonials-widget.php
@@ -123,6 +123,17 @@ class Testimonials_Widget extends Aihrus_Common {
 		add_filter( 'plugin_row_meta', array( __CLASS__, 'plugin_row_meta' ), 10, 2 );
 		add_filter( 'post_updated_messages', array( __CLASS__, 'post_updated_messages' ) );
 		add_filter( 'pre_get_posts', array( __CLASS__, 'pre_get_posts_author' ) );
+
+		if ( self::do_load() ) {
+			add_filter( 'manage_category_custom_column', array( __CLASS__, 'category_column' ), 10, 3 );
+			add_filter( 'manage_edit-category_columns', array( __CLASS__, 'category_columns' ) );
+			add_filter( 'manage_edit-post_tag_columns', array( __CLASS__, 'category_columns' ) );
+			add_filter( 'manage_edit-testimonials-widget-category_columns', array( __CLASS__, 'category_columns' ) );
+			add_filter( 'manage_edit-testimonials-widget-post_tag_columns', array( __CLASS__, 'category_columns' ) );
+			add_filter( 'manage_post_tag_custom_column', array( __CLASS__, 'post_tag_column' ), 10, 3 );
+			add_filter( 'manage_testimonials-widget-category_custom_column', array( __CLASS__, 'category_column' ), 10, 3 );
+			add_filter( 'manage_testimonials-widget-post_tag_custom_column', array( __CLASS__, 'post_tag_column' ), 10, 3 );
+		}
 	}
 
 
@@ -132,7 +143,7 @@ class Testimonials_Widget extends Aihrus_Common {
 
 
 	public static function init() {
-		add_filter( 'the_content', array( __CLASS__, 'get_single' ) );
+		add_filter( 'the_content', array( __CLASS__, 'get_single' ), -1 );
 
 		load_plugin_textdomain( self::PT, false, 'testimonials-widget/languages' );
 
@@ -140,10 +151,7 @@ class Testimonials_Widget extends Aihrus_Common {
 		self::$cpt_tags     = self::PT . '-post_tag';
 
 		self::init_post_type();
-
-		$force_css_loading = tw_get_option( 'force_css_loading' );
-		if ( $force_css_loading )
-			self::styles();
+		self::styles();
 	}
 
 
@@ -216,16 +224,9 @@ class Testimonials_Widget extends Aihrus_Common {
 			$details = self::get_testimonial_html( $testimonial, $atts );
 			$details = apply_filters( 'testimonials_widget_testimonial_html_single', $details, $testimonial, $atts );
 
-			$do_schema = $atts['enable_schema'];
-			if ( $do_schema )
-				$content = self::create_schema_span( self::$review_body, $content );
-
 			$content = apply_filters( 'testimonials_widget_testimonial_html_single_content', $content, $testimonial, $atts );
 
 			$text = $content . $details;
-			if ( $do_schema )
-				$text = self::create_schema_div_prop( self::$cw_review, self::$review_schema, $text );
-
 			$text = apply_filters( 'testimonials_widget_cache_set', $text, $atts );
 		}
 
@@ -330,7 +331,7 @@ class Testimonials_Widget extends Aihrus_Common {
 	public static function notice_2_12_0() {
 		$text = sprintf( __( 'If your Testimonials display has gone to funky town, please <a href="%s">read the FAQ</a> about possible CSS fixes.', 'testimonials-widget' ), esc_url( 'https://aihrus.zendesk.com/entries/23722573-Major-Changes-Since-2-10-0' ) );
 
-		self::notice_updated( $text );
+		aihr_notice_updated( $text );
 	}
 
 
@@ -586,6 +587,7 @@ class Testimonials_Widget extends Aihrus_Common {
 			'title',
 			'editor',
 			'thumbnail',
+			'publicize',
 		);
 
 		$allow_comments = tw_get_option( 'allow_comments', false );
@@ -742,11 +744,11 @@ class Testimonials_Widget extends Aihrus_Common {
 		if ( $use_bxslider ) {
 			$enable_video = $atts['enable_video'];
 			if ( $enable_video ) {
-				wp_register_script( 'jquery.fitvids', self::$plugin_assets . 'js/jquery.fitvids.js', array( 'jquery' ), '1.0' );
+				wp_register_script( 'jquery.fitvids', self::$plugin_assets . 'js/jquery.fitvids.js', array( 'jquery' ), '1.0', true );
 				wp_enqueue_script( 'jquery.fitvids' );
 			}
 
-			wp_register_script( 'jquery.bxslider', self::$plugin_assets . 'js/jquery.bxslider.js', array( 'jquery' ), '4.1.1' );
+			wp_register_script( 'jquery.bxslider', self::$plugin_assets . 'js/jquery.bxslider.js', array( 'jquery' ), '4.1.1', true );
 			wp_enqueue_script( 'jquery.bxslider' );
 		}
 
@@ -979,6 +981,7 @@ EOF;
 			$div_open = '<div class="' . $id . ' ' . $id_base . '">';
 		}
 
+		$div_open .= "\n";
 		if ( empty( $testimonials ) && ! $hide_not_found ) {
 			$testimonials = array(
 				array( 'testimonial_content' => esc_html__( 'No testimonials found', 'testimonials-widget' ) ),
@@ -1009,7 +1012,8 @@ EOF;
 		if ( $paging || $paging_after )
 			$post_paging = self::get_testimonials_paging( $atts, false );
 
-		$div_close = '</div>';
+		$div_close  = '</div>';
+		$div_close .= "\n";
 
 		$html = $div_open
 			. $pre_paging
@@ -1027,7 +1031,6 @@ EOF;
 		$disable_quotes  = $atts['disable_quotes'];
 		$do_image        = ! $atts['hide_image'] && ! empty( $testimonial['testimonial_image'] );
 		$do_image_single = ! $atts['hide_image_single'];
-		$do_content      = ! $atts['hide_content'];
 		$do_schema       = $atts['enable_schema'];
 		$keep_whitespace = $atts['keep_whitespace'];
 		$remove_hentry   = $atts['remove_hentry'];
@@ -1058,8 +1061,10 @@ EOF;
 		$div_open  = '<div class="' . $class . '">';
 		$div_open .= '<!-- ' . self::ID . ":{$post_id}: -->";
 
-		if ( $do_schema && $do_content )
+		if ( $do_schema ) {
+			$div_open .= "\n";
 			$div_open .= sprintf( self::$schema_div_open, self::$review_schema );
+		}
 
 		if ( $remove_hentry )
 			$div_open = str_replace( ' hentry', '', $div_open );
@@ -1093,6 +1098,7 @@ EOF;
 			$extra .= '<div class="extra">';
 			$extra .= $testimonial['testimonial_extra'];
 			$extra .= '</div>';
+			$extra .= "\n";
 		}
 
 		$bottom_text = '';
@@ -1100,13 +1106,17 @@ EOF;
 			$bottom_text  = '<div class="bottom_text">';
 			$bottom_text .= $atts['bottom_text'];
 			$bottom_text .= '</div>';
+			$bottom_text .= "\n";
 		}
 
 		$div_close = '';
-		if ( $do_schema && $do_content )
+		if ( $do_schema ) {
 			$div_close .= '</div>';
+			$div_close .= "\n";
+		}
 
 		$div_close .= '</div>';
+		$div_close .= "\n";
 
 		$html = $div_open
 			. $image
@@ -1133,17 +1143,12 @@ EOF;
 		$content_more  = apply_filters( 'testimonials_widget_content_more', esc_html__( '…', 'testimonials-widget' ) );
 		$content_more .= self::$tag_close_quote;
 		$do_content    = ! $atts['hide_content'] && ! empty( $testimonial['testimonial_content'] );
-		$do_schema     = $atts['enable_schema'];
 		$use_quote_tag = $atts['use_quote_tag'];
 
 		$quote = '';
 		if ( $do_content ) {
 			$content = $testimonial['testimonial_content'];
-			if ( $do_schema )
-				$content = self::create_schema_span( self::$review_body, $content );
-
 			$content = self::format_content( $content, $widget_number, $atts );
-
 			if ( $char_limit ) {
 				$content = self::testimonials_truncate( $content, $char_limit, $content_more );
 				$content = force_balance_tags( $content );
@@ -1161,6 +1166,8 @@ EOF;
 				$quote .= $content;
 				$quote .= '</q>';
 			}
+			
+			$quote = "\n" . $quote;
 		}
 
 		return $quote;
@@ -1250,7 +1257,7 @@ EOF;
 				$temp .= $cite;
 				$temp .= '</div>';
 
-				$cite = $temp;
+				$cite = "\n" . $temp . "\n";
 			} else {
 				$cite = '<cite>' . $cite . '</cite>';
 			}
@@ -1329,6 +1336,7 @@ EOF;
 			$html .= ' append';
 
 		$html .= '">';
+		$html .= "\n";
 
 		if ( $atts['paged'] )
 			$paged = $atts['paged'];
@@ -1336,23 +1344,23 @@ EOF;
 			$paged = 1;
 
 		if ( ! function_exists( 'wp_pagenavi' ) ) {
-			$html .= '	<div class="alignleft">';
+			$html .= '<div class="alignleft">';
 
 			if ( 1 < $paged ) {
 				$laquo = apply_filters( 'testimonials_widget_previous_posts_link_text', esc_html__( '&laquo;', 'testimonials-widget' ) );
 				$html .= get_previous_posts_link( $laquo, $paged );
 			}
 
-			$html .= '	</div>';
-
-			$html .= '	<div class="alignright">';
-
+			$html .= '</div>';
+			$html .= "\n";
+			$html .= '<div class="alignright">';
 			if ( $paged != self::$max_num_pages ) {
 				$raquo = apply_filters( 'testimonials_widget_next_posts_link_text', esc_html__( '&raquo;', 'testimonials-widget' ) );
 				$html .= get_next_posts_link( $raquo, self::$max_num_pages );
 			}
 
-			$html .= '	</div>';
+			$html .= '</div>';
+			$html .= "\n";
 		} else {
 			$args = array(
 				'echo' => false,
@@ -1361,9 +1369,11 @@ EOF;
 			$args = apply_filters( 'testimonials_widget_wp_pagenavi', $args );
 
 			$html .= wp_pagenavi( $args );
+			$html .= "\n";
 		}
 
 		$html .= '</div>';
+		$html .= "\n";
 
 		return $html;
 	}
@@ -1747,6 +1757,7 @@ EOF;
 		extract( $testimonial );
 
 		$do_company  = ! $atts['hide_company'] && ! empty( $testimonial_company );
+		$do_content  = ! empty( $testimonial['testimonial_content'] );
 		$do_email    = ! $atts['hide_email'] && ! empty( $testimonial_email ) && is_email( $testimonial_email );
 		$do_image    = ! $atts['hide_image'] && ! empty( $testimonial_image );
 		$do_location = ! $atts['hide_location'] && ! empty( $testimonial_location );
@@ -1757,7 +1768,8 @@ EOF;
 		$item_reviewed     = self::clean_string( $atts['item_reviewed'] );
 		$item_reviewed_url = self::clean_string( $atts['item_reviewed_url'] );
 
-		$schema = '';
+		$schema  = '<div style="display: none;">';
+		$schema .= "\n";
 
 		$agg_meta      = array();
 		$author_meta   = array();
@@ -1802,12 +1814,16 @@ EOF;
 		$author_meta = apply_filters( 'testimonials_widget_schema_author', $author_meta, $testimonial, $atts );
 		$author      = self::create_schema_div_prop( self::$cw_author, self::$person_schema, $author_meta );
 		$schema     .= $author;
+		$schema     .= "\n";
 
 		$post         = get_post( $testimonial['post_id'] );
 		$the_date     = mysql2date( 'Y-m-d', $post->post_date );
 		$the_date_mod = mysql2date( 'Y-m-d', $post->post_modified );
 
 		$review_name_length = apply_filters( 'testimonials_widget_review_name_length', 156 );
+
+		if ( $do_content )
+			$review_meta[ self::$review_body ] = $testimonial['testimonial_content'];
 
 		$review_meta[ self::$cw_date ]     = $the_date;
 		$review_meta[ self::$cw_date_mod ] = $the_date_mod;
@@ -1823,12 +1839,14 @@ EOF;
 		$review_meta = apply_filters( 'testimonials_widget_schema_review', $review_meta, $testimonial, $atts );
 		$review      = self::create_schema_meta( $review_meta );
 		$schema     .= $review;
+		$schema     .= "\n";
 
 		$agg_meta[ self::$agg_count ] = self::$found_posts;
 
 		$agg_meta  = apply_filters( 'testimonials_widget_schema_aggregate', $agg_meta, $testimonial, $atts );
 		$aggregate = self::create_schema_div_prop( self::$cw_aggregate, self::$agg_schema, $agg_meta );
 		$schema   .= $aggregate;
+		$schema   .= "\n";
 
 		$item_meta[ self::$thing_name ] = $item_reviewed;
 		$item_meta[ self::$thing_url ]  = $item_reviewed_url;
@@ -1836,6 +1854,10 @@ EOF;
 		$item_meta = apply_filters( 'testimonials_widget_schema_item', $item_meta, $testimonial, $atts );
 		$item      = self::create_schema_div_prop( self::$review_item, self::$thing_schema, $item_meta );
 		$schema   .= $item;
+		$schema   .= "\n";
+
+		$schema   .= '</div>';
+		$schema   .= "\n";
 
 		$schema = apply_filters( 'testimonials_widget_schema', $schema, $testimonial, $atts );
 
@@ -1854,6 +1876,8 @@ EOF;
 				$meta .= self::create_schema_div_prop( $key, $value[ 0 ], $value[ 1 ] );
 			else
 				$meta .= sprintf( self::$schema_meta, $key, $value );
+
+			$meta .= "\n";
 		}
 
 		return $meta;
@@ -1885,6 +1909,8 @@ EOF;
 					$meta .= self::create_schema_div_prop( $key, $value[ 0 ], $value[ 1 ] );
 				else
 					$meta .= sprintf( self::$schema_meta, $key, $value );
+				
+				$meta .= "\n";
 			}
 
 			$schema = sprintf( self::$schema_div_prop, $property_name, $schema_name, $meta );
@@ -1932,7 +1958,6 @@ EOF;
 			$widget_number = Testimonials_Widget::get_instance();
 
 		self::scripts( $atts );
-		self::styles();
 
 		self::generate_css( $atts );
 		self::generate_js( $testimonials, $atts, $widget_number );
@@ -1942,7 +1967,7 @@ EOF;
 	public static function make_gravatar_featured( $post_id, $email ) {
 		$size  = get_option( 'large_size_w' );
 		$image = get_avatar( $email, $size );
-		$src   = self::get_image_src( $image );
+		$src   = self::get_image_src( $image, false );
 		$file  = sanitize_title( $email ) . '.jpeg';
 
 		self::add_media( $post_id, $src, $file );
@@ -1952,7 +1977,7 @@ EOF;
 	public static function notice_2_15_0() {
 		$text = sprintf( __( 'If your Testimonials display has gone to funky town, please <a href="%s">read the FAQ</a> about possible fixes.', 'testimonials-widget' ), esc_url( 'https://aihrus.zendesk.com/entries/28402246-Major-Change-for-2-15-0' ) );
 
-		self::notice_updated( $text );
+		aihr_notice_updated( $text );
 	}
 
 
@@ -2019,7 +2044,7 @@ EOD;
 		if ( ! empty( $GLOBALS['pagenow'] ) && in_array( $GLOBALS['pagenow'], array( 'options.php' ) ) ) {
 			$do_load = true;
 		} elseif ( ! empty( $_REQUEST['post_type'] ) && self::PT == $_REQUEST['post_type'] ) {
-			if ( ! empty( $GLOBALS['pagenow'] ) && in_array( $GLOBALS['pagenow'], array( 'edit.php' ) ) ) {
+			if ( ! empty( $GLOBALS['pagenow'] ) && in_array( $GLOBALS['pagenow'], array( 'edit.php', 'edit-tags.php' ) ) ) {
 				$do_load = true;
 			} elseif ( ! empty( $_REQUEST['page'] ) && Testimonials_Widget_Settings::ID == $_REQUEST['page'] ) {
 				$do_load = true;
@@ -2028,6 +2053,47 @@ EOD;
 
 		return $do_load;
 	}
+
+
+	public static function category_columns( $columns ) {
+		$columns['shortcode'] = esc_html__( 'Shortcode', 'testimonials-widget' );
+
+		return $columns;
+	}
+
+
+	public static function category_column( $result, $column_name, $term_id, $category = true ) {
+		$attribute = $category ? 'category' : 'tags';
+
+		$use_cpt_taxonomy = tw_get_option( 'use_cpt_taxonomy', false );
+		if ( ! $use_cpt_taxonomy ) {
+			if ( $category )
+				$term = get_term( $term_id, 'category' );
+			else
+				$term = get_term( $term_id, 'post_tag' );
+		} else {
+			if ( $category )
+				$term = get_term( $term_id, self::$cpt_category );
+			else
+				$term = get_term( $term_id, self::$cpt_tags );
+		}
+
+		switch ( $column_name ) {
+			case 'shortcode':	
+				$result  = '[testimonialswidget_list ' . $attribute . '="' .$term->slug . '"]';
+				$result .= '<br />';
+				$result .= '[testimonialswidget_widget ' . $attribute . '="' .$term->slug . '"]';
+				break;
+		}
+
+		return $result;	
+	}
+
+
+	public static function post_tag_column( $result, $column_name, $term_id ) {
+		return self::category_column( $result, $column_name, $term_id, false );
+	}
 }
+
 
 ?>